fastmcp_template/service.py
from fastapi import FastAPI, HTTPException
from pydantic import BaseModel
from typing import Optional, Dict, Any, List
from langgraph.graph import StateGraph, START, END
from langgraph.checkpoint.memory import MemorySaver
from langgraph.types import Command
from dotenv import load_dotenv
from utils.utils import write_to_log
import os

# Load environment variables
load_dotenv()

app = FastAPI()

# Initialize in-memory state for conversations
conversation_memory = {}

class AgentState(TypedDict):
    latest_user_message: str
    messages: List[bytes]
    scope: str
    documentation: Dict[str, Any]
    implementation_plan: str

class InvokeRequest(BaseModel):
    message: str
    thread_id: str
    doc_url: Optional[str] = None
    is_first_message: bool = False
    config: Optional[Dict[str, Any]] = None

@app.get("/health")
async def health_check():
    """Health check endpoint"""
    return {"status": "ok"}    

async def process_documentation(doc_url: str) -> Dict[str, Any]:
    """Process API documentation from URL."""
    # TODO: Implement documentation processing
    # This should handle different formats (OpenAPI, Markdown, etc.)
    return {"url": doc_url, "content": "Processed documentation"}

async def define_scope(state: AgentState) -> Dict[str, Any]:
    """Define the scope of the MCP implementation."""
    prompt = f"""
    User Request: {state['latest_user_message']}
    Documentation: {state['documentation']}
    
    Create detailed scope for the MCP implementation including:
    - Core components needed
    - External dependencies
    - Integration points
    - Testing strategy
    """
    # TODO: Use LLM to generate scope
    return {"scope": "Generated scope"}

async def create_implementation_plan(state: AgentState) -> Dict[str, Any]:
    """Create implementation plan based on scope."""
    prompt = f"""
    Scope: {state['scope']}
    Documentation: {state['documentation']}
    
    Create detailed implementation plan including:
    - File structure
    - Core functions needed
    - Integration approach
    - Testing plan
    """
    # TODO: Use LLM to generate plan
    return {"implementation_plan": "Generated plan"}

async def generate_code(state: AgentState) -> Dict[str, Any]:
    """Generate MCP implementation code."""
    # TODO: Use LLM to generate code
    return {"code": "Generated code"}

# Build workflow
builder = StateGraph(AgentState)

# Add nodes
builder.add_node("process_documentation", process_documentation)
builder.add_node("define_scope", define_scope)
builder.add_node("create_implementation_plan", create_implementation_plan)
builder.add_node("generate_code", generate_code)

# Set edges
builder.add_edge(START, "process_documentation")
builder.add_edge("process_documentation", "define_scope")
builder.add_edge("define_scope", "create_implementation_plan")
builder.add_edge("create_implementation_plan", "generate_code")
builder.add_edge("generate_code", END)

# Configure persistence
memory = MemorySaver()
workflow = builder.compile(checkpointer=memory)

@app.post("/invoke")
async def invoke(request: InvokeRequest):
    try:
        if request.is_first_message and not request.doc_url:
            raise HTTPException(status_code=400, detail="Documentation URL required for first message")
        
        # Initialize or update state
        state = {
            "latest_user_message": request.message,
            "messages": [],
            "documentation": await process_documentation(request.doc_url) if request.doc_url else {},
            "scope": "",
            "implementation_plan": ""
        }
        
        # Run workflow
        result = await workflow.ainvoke(state)
        return {"response": result}
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8100) 