fastmcp_template/requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
httpx
pyyaml
beautifulsoup4
markdown
openai
pydantic
fastmcp 